src/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pawlaczyk"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.12.0"
django-filter = "^21.1"
django-extensions = "^3.1.5"
requests = "^2.27.1"
pymongo = "^4.1.1"
pandas = "^1.4.2"
eventlet = "^0.33.1"
python-socketio = "5.5.2"
selenium = "^4.2.0"
beautifulsoup4 = "^4.11.1"
xmltodict = "^0.13.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
